Simplify build reqs and use more macros (thanks @ignatenkobrain)
This commit simplifies some of the build requirements for the package.
It also uses the full system path for some packages to make it easier to
look up binaries that use different packages or if the binary moves
packages.

Additionally, this commit also adds more macros to successfully build
the package for Fedora and EPEL, thanks to better use of RPM macros.

Most of the changes here were suggested by @ignatenkobrain – thanks!