mpris-scrobbler: install systemd on F29 or older
A special package, `systemd-rpm-macros`, is available on Fedora 30 or
later. Previous versions of Fedora need the full `systemd` package as a
build requirement for this package to build successfully. This commit
adds an additional check for version of Fedora so this package builds
for F29 or older versions of Fedora.