:boom: Drop Dockerfiles in favor of mock
When I created these Dockerfiles, it was because I was new to packaging
and containers were easier for me to test the package builds than with
local tools. Now that I am more familiar with the Fedora packaging
toolset and RPM, it makes more sense to just use `mock` in most cases,
or use the provided Fedora build infrastructure where applicable.

:boom:

Signed-off-by: Justin W. Flory <git@jwf.io>